Consolidated models to a model.py file Moved some view functionality to utilities.py to avoid repeating code began checkout.html
