--- TeamIGS/Customer.py
@@ -1,11 +1,0 @@
-# Handles User information and account management
-from django.db import models
-from django.contrib.auth.models import User
-
-class Customer(models.Model):
-	user = models.OneToOneField(User, null=True, blank=True, on_delete=models.CASCADE)
-	name = models.CharField(max_length=200, null=True)
-	email = models.CharField(max_length=200)
-
-	def __str__(self):
-		return self.name
