--- TeamIGS/Order.py
@@ -1,29 +1,0 @@
-# Handles processing of order information and displaying it to the user
-# Might serve same function as cart, requires more development to be sure
-from django.db import models
-from .Item import Item
-from .Customer import Customer
-import datetime
-from django.conf import settings
-
-
-class Order(models.Model):
-	customer = models.ForeignKey(Customer, on_delete=models.SET_NULL, null=True, blank=True)
-	date_ordered = models.DateTimeField(auto_now_add=True)
-	complete = models.BooleanField(default=False)
-	transaction_id = models.CharField(max_length=100, null=True)
-
-	def __str__(self):
-		return str(self.id)
-
-	@property
-	def get_cart_total(self):
-		orderitems = self.orderitem_set.all()
-		total = sum([item.get_total for item in orderitems])
-		return total 
-
-	@property
-	def get_cart_items(self):
-		orderitems = self.orderitem_set.all()
-		total = sum([item.quantity for item in orderitems])
-		return total
