--- TeamIGS/Shipping.py
@@ -1,9 +1,0 @@
-#Handles processing related to shipping, including cost and time calculations as well as tracking
-class Shipping():
-
-    def __init__(self):
-        pass
-
-    # Calculates cost of shipping an item
-    def calculate_cost(self, weight, distance):
-        pass
