+++ TeamIGS/templates/TeamIGS/checkout.html
@@ -1,16 +1,16 @@
-<div class="box-element">
-	<a  class="btn btn-outline-dark" href="{% url 'cart' %}">&#x2190; Back to Cart</a>
-	<hr>
-	<h3>Order Summary</h3>
-	<hr>
-	{% for item in items %}
-	<div class="cart-row">
-		<div style="flex:2"><img class="row-image" src="{{item.product.imageURL}}"></div>
-		<div style="flex:2"><p>{{item.product.name}}</p></div>
-		<div style="flex:1"><p>${{item.product.price|floatformat:2}}</p></div>
-		<div style="flex:1"><p>x{{item.quantity}}</p></div>
+{% load static %}
+{% block content %}
+	<div class="row">
+		<div class="col-lg-6">
+			<div class="box-element" id="form-wrapper">
+
+			</div>
+
+		</div>
+
+		<div class="col-log-6">
+			<div class="box-element">
+
+			</div>
+		</div>
 	</div>
-	{% endfor %}
-	<h5>Items:   {{order.get_cart_items}}</h5>
-	<h5>Total:   ${{order.get_cart_total|floatformat:2}}</h5>
-</div>
