+++ TeamIGS/urls.py
@@ -20,8 +20,6 @@ from .views import (
     IndexView,
     DetailView,
     cart,
-    InCategoryView,
-    loginView,
     checkout,
     updateItem,
 )
@@ -37,12 +35,13 @@ urlpatterns = [
     path("cart/", cart, name="cart"),
     path("checkout/", checkout, name="checkout"),
     path("updateItem/", updateItem, name="updateItem"),
-    # Category pages moved to low priority, might not be properly implemented for a while
+    
+    # Not implemented yet
     # Page containing list of categories, not working
-    path("category/", InCategoryView.as_view(), name="category"),
+    # path("category/", InCategoryView.as_view(), name="category"),
     # Page containing items within category
-    path("category/<str:name>/", InCategoryView.as_view(), name="inCategory"),
+    # path("category/<str:name>/", InCategoryView.as_view(), name="inCategory"),
     # Currently placeholder, requires login.html to be made
-    path("login/", loginView.as_view(), name="login"),
+    # path("login/", loginView.as_view(), name="login"),
 ] 
 urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
