+++ TeamIGS/views.py
@@ -12,12 +12,9 @@ from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login
 from django.http import JsonResponse
-from .Item import Item
-from .Order import Order
-from .Category import Category
-from .OrderItem import OrderItem
-from .Customer import Customer
+from .models import Category, Customer, Item, Order, OrderItem
 import json
+from .utilities import cartFromCookie
 
 class IndexView(generic.ListView):
     template_name = "TeamIGS/index.html"
@@ -31,39 +28,20 @@ class DetailView(generic.DetailView):
     model = Item
     template_name = "TeamIGS/detail.html"
 
-
-# # Requires updated implementation
-# class CartView(generic.ListView):
-#     template_name = "TeamIGS/cart.html"
-#     context_object_name = "cart_items"
-
-#     def get_queryset(self):
-#         return Item.objects.filter() # Doesn't filter, fix later when rest is working
-
-# Requires more implementation, currently always kicks user back to index but should visit page when
-# user has items in cart
-# class CartView(LoginRequiredMixin, View):
-#     def get(self, *args, **kwargs):
-#         try:
-#             order = Order.objects.get(user=self.request.user, ordered=False)
-#             context = {'object': order}
-#             return render(self.request, 'TeamIGS/cart.html', context)
-#         except ObjectDoesNotExist:
-#             # messages.error(self.request, "There is no active order")
-#             return redirect("/")
-
 def checkout(request):
-	if request.user.is_authenticated:
-		customer = request.user.customer
-		order, created = Order.objects.get_or_create(customer=customer, complete=False)
-		items = order.orderitem_set.all()
-	else:
-		#Create empty cart for now for non-logged in user
-		items = []
-		order = {'get_cart_total':0, 'get_cart_items':0}
+    # Commented out until accounts are implemented to prevent bugs
+    # if request.user.is_authenticated:
+    #     customer = request.user.customer
+    #     order, created = Order.objects.get_or_create(customer=customer, complete=False)
+    #     items = order.orderitem_set.all()
+    # else:
+    cartData = cartFromCookie(request)
+    cartItems = cartData['cartItems']
+    order = cartData['order']
+    items = cartData['items']
 
-	context = {'items':items, 'order':order}
-	return render(request, 'TeamIGS/checkout.html', context)
+    context = {'items':items, 'order':order}
+    return render(request, 'TeamIGS/checkout.html', context)
 
 def cart(request):
     # Removing Temporarily until accounts are implemented
@@ -73,34 +51,15 @@ def cart(request):
     #     order, created = Order.objects.get_or_create(customer=customer, complete=False)
     #     items = order.orderitem_set.all()
     # else:
-    cart = json.loads(request.COOKIES['cart'])
-    print("Cart:", cart)
-    items = []
-    order = {'get_cart_total':0, 'get_cart_items':0}
-    cartItems = order['get_cart_items']
+    cartData = cartFromCookie(request)
+    cartItems = cartData['cartItems']
+    order = cartData['order']
+    items = cartData['items']
+
     
-    for i in cart:
-        cartItems += cart[i]['quantity']
-        item = Item.objects.get(id=i)
-        total = (item.price * cart[i]['quantity'])
-
-        order['get_cart_total'] += total
-        order['get_cart_items'] += cart[i]['quantity']
-
-        item = {
-            'item':{
-                'id': item.id,
-                'name': item.name,
-                'price': item.price,
-                'image': item.image,
-            },
-            'quantity': cart[i]['quantity'],
-            'get_total':total,
-        }
-        items.append(item)
-
-    context = {'items':items, 'order':order}
+    context = {'items':items, 'order':order, 'cartItems':cartItems}
     return render(request, 'TeamIGS/cart.html/', context)
+        
 
 def updateItem(request):
 	data = json.loads(request.body)
@@ -135,46 +94,46 @@ def updateItem(request):
 #     def get_queryset(self):
 #         return Category.objects.order_by("name")
 
-class InCategoryView(generic.ListView):
-    template_name = "TeamIGS/InCategory.html"
-    context_object_name = "category_items"
+# class InCategoryView(generic.ListView):
+#     template_name = "TeamIGS/InCategory.html"
+#     context_object_name = "category_items"
 
-    def get_queryset(self):
-        categoryID = self.request.GET.get('category')
-        if categoryID:
-            return Item.objects.filter(category__name=categoryID)
-        else:  
-            return
+#     def get_queryset(self):
+#         categoryID = self.request.GET.get('category')
+#         if categoryID:
+#             return Item.objects.filter(category__name=categoryID)
+#         else:  
+#             return
 
-# Hasn't been tested
-# Might want to add email verification in the future, though that might be handled elsewhere
-def createAccount(request):
-    username = request.POST["username"]
-    email = request.POST["email"]
-    password = request.POST["password"]
-    firstName = request.POST["firstname"]
-    lastName = request.POST["lastname"]
-    user = User.objects.create_user(username, email, password)
-    user.first_name = firstName
-    user.last_name = lastName
-    messages.add_message(self.request, "Account Created!")
-    return render(self.request, 'TeamIGS/')
+# # Hasn't been tested
+# # Might want to add email verification in the future, though that might be handled elsewhere
+# def createAccount(request):
+#     username = request.POST["username"]
+#     email = request.POST["email"]
+#     password = request.POST["password"]
+#     firstName = request.POST["firstname"]
+#     lastName = request.POST["lastname"]
+#     user = User.objects.create_user(username, email, password)
+#     user.first_name = firstName
+#     user.last_name = lastName
+#     messages.add_message(self.request, "Account Created!")
+#     return render(self.request, 'TeamIGS/')
 
-# Directed to this page when attempting to sign in
-# Verifies account and sends to home or returns error and sends back to login page
-class loginView(generic.View):
-    template_name = "TeamIGS/login.html"
+# # Directed to this page when attempting to sign in
+# # Verifies account and sends to home or returns error and sends back to login page
+# class loginView(generic.View):
+#     template_name = "TeamIGS/login.html"
 
-    def loginView(request):
-        username = request.POST["username"]
-        password = request.POST["password"]
-        user = authenticate(request, username, password)
-        if user is not None:
-            login(request, user)
-            # Can make success page for login, otherwise this works
-            messages.add_message(self.request, "Successful login")
-            return render(self.request, 'TeamIGS/index.html')
-        else:
-            # Make failure page
-            messages.error(self.request, 'Failed to login')
-            return render(self.request, 'TeamIGS/login.html')
+#     def loginView(request):
+#         username = request.POST["username"]
+#         password = request.POST["password"]
+#         user = authenticate(request, username, password)
+#         if user is not None:
+#             login(request, user)
+#             # Can make success page for login, otherwise this works
+#             messages.add_message(self.request, "Successful login")
+#             return render(self.request, 'TeamIGS/index.html')
+#         else:
+#             # Make failure page
+#             messages.error(self.request, 'Failed to login')
+#             return render(self.request, 'TeamIGS/login.html')
